[AICOMRCCL-1934] [test] Sort dev_runtime stubs into shared fakes layer
Move the ~40 dev_runtime-only symbols out of the target-local
DevRuntimeTestsStubs.cc and into the reusable test/host/fakes/ layer so any
other host-only micro-test can link them:

- Generic nccl/bootstrap/argcheck/group stubs (ncclGetErrorString,
  bootstrapAllGather/Barrier/IntraNode*, PtrCheck, CommCheck,
  ncclCommEnsureReady, ncclCommRegister/Deregister, ncclCommWindowDeregister,
  ncclGroupStartInternal/EndInternal) -> nccl_fakes.cc.

- New fakes/devr_fakes.{h,cc} (device-runtime layer) for the devr / gin / rma /
  team / shadow-pool / space-allocator / symmetric-kernel symbols, the
  devcomm compat tables, and ncclMemoryStack::allocateSpilled.

DevRuntimeTestsStubs.cc now holds only what is genuinely specific to that
binary: the thread-local group globals and the host-memory-backed HIP VMM/
stream fakes (shrinks from 296 to 131 lines). Wire devr_fakes.cc into the
rccl-UnitTestsDevRuntime target.

Verified: devr_fakes.cc, nccl_fakes.cc and the trimmed DevRuntimeTestsStubs.cc
all compile against the hipified DevRuntime include set; rccl-UnitTestsMicro
(which links nccl_fakes.cc) still builds and all 30 tests pass.
